fix: Kotlin compilation errors — Keccak constants, RLP bytes, BC API compatibility

+++ native/android/tech/atlasnexus/ai2work/wallet/RLP.kt
@@ -6,9 +6,9 @@ import java.math.BigInteger
 /** Minimal RLP encoder for Ethereum transactions. */
 object RLP {
     fun bytes(data: ByteArray): ByteArray {
-        if (data.isEmpty()) return byteArrayOf(0x80)
+        if (data.isEmpty()) return byteArrayOf(0x80.toByte())
         if (data.size == 1 && (data[0].toInt() and 0xFF) < 0x80) return data
-        return lenPrefix(data.size, 0x80) + data
+        return lenPrefix(data.size, 0x80.toByte()) + data
     }
 
     fun long(v: Long): ByteArray = bigint(BigInteger.valueOf(v))
